Moved mappings of redmineModels out of DataService into RedmineMapper.
